+++ collective/elasticsearch/es.py
@@ -1,11 +1,12 @@
 # -*- coding: utf-8 -*-
-from collective.elasticsearch import hook
-from collective.elasticsearch import logger
+import math
+
+from collective.elasticsearch import hook, logger
 from collective.elasticsearch.brain import BrainFactory
-from collective.elasticsearch.interfaces import IElasticSearchCatalog
-from collective.elasticsearch.interfaces import IElasticSettings
-from collective.elasticsearch.interfaces import IMappingProvider
-from collective.elasticsearch.interfaces import IQueryAssembler
+from collective.elasticsearch.interfaces import (IElasticSearchCatalog,
+                                                 IElasticSettings,
+                                                 IMappingProvider,
+                                                 IQueryAssembler)
 from collective.elasticsearch.utils import getESOnlyIndexes
 from DateTime import DateTime
 from elasticsearch import Elasticsearch
@@ -13,31 +14,15 @@ from elasticsearch.exceptions import NotFoundError
 from plone import api
 from plone.registry.interfaces import IRegistry
 from Products.CMFCore.permissions import AccessInactivePortalContent
-from Products.CMFCore.utils import _checkPermission
-from Products.CMFCore.utils import _getAuthenticatedUser
+from Products.CMFCore.utils import _checkPermission, _getAuthenticatedUser
 from Products.ZCatalog.Lazy import LazyMap
-from zope.component import ComponentLookupError
-from zope.component import getMultiAdapter
-from zope.component import getUtility
+from zope.component import ComponentLookupError, getMultiAdapter, getUtility
 from zope.globalrequest import getRequest
 from zope.interface import implementer
-
 
 CONVERTED_ATTR = '_elasticconverted'
 CUSTOM_INDEX_NAME_ATTR = '_elasticcustomindex'
 INDEX_VERSION_ATTR = '_elasticindexversion'
-
-
-def _reversed_sort(sort):
-    criterion, order = sort.split(':')
-    if order == 'desc':
-        order = 'asc'
-    elif order == 'asc':
-        order = 'desc'
-    else:
-        raise ValueError("Invalid order: %s" % order)
-
-    return "%s:%s" % (criterion, order)
 
 
 class ElasticResult(object):
@@ -48,23 +33,43 @@ class ElasticResult(object):
         self.es = es
         self.bulk_size = es.get_setting('bulk_size', 50)
         qassembler = getMultiAdapter((getRequest(), es), IQueryAssembler)
-        dquery, sort = qassembler.normalize(query)
-        equery = qassembler(dquery)
+        dquery, self.sort = qassembler.normalize(query)
+        self.query = qassembler(dquery)
 
         # results are stored in a dictionary, keyed
         # but the start index of the bulk size for the
         # results it holds. This way we can skip around
         # for result data in a result object
-        self.query = equery
-        result = es._search(self.query, sort=sort, **query_params)['hits']
+        result = es._search(self.query, sort=self.sort, **query_params)['hits']
         self.results = {
             0: result['hits']
         }
         self.count = result['total']
-        self.sort = sort
         self.query_params = query_params
 
+    def __len__(self):
+        return self.count
+
     def __getitem__(self, key):
+        '''
+        Lazy loading es results with negative index support.
+
+        We store the results in buckets of what the bulk size is.
+
+        This is so you can skip around in the indexes without needing
+        to load all the data.
+
+        Example(all zero based indexing here remember):
+            (525 results with bulk size 50)
+            - self[0]: 0 bucket, 0 item
+            - self[10]: 0 bucket, 10 item
+            - self[50]: 50 bucket: 0 item
+            - self[55]: 50 bucket: 5 item
+            - self[352]: 350 bucket: 2 item
+            - self[-1]: 500 bucket: 24 item
+            - self[-2]: 500 bucket: 23 item
+            - self[-55]: 450 bucket: 19 item
+        '''
         if isinstance(key, slice):
             return [self[i] for i in range(key.start, key.end)]
         else:
@@ -72,21 +77,28 @@ class ElasticResult(object):
                 raise IndexError
             elif key < 0 and abs(key) > self.count:
                 raise IndexError
-            elif key >= 0:
+
+            if key >= 0:
                 result_key = (key / self.bulk_size) * self.bulk_size
                 start = result_key
-                sort = self.sort
                 result_index = key % self.bulk_size
             elif key < 0:
-                result_key = - ((abs(key) / self.bulk_size) * self.bulk_size) - 1  # noqa
-                start = abs(result_key) - 1
-                sort = _reversed_sort(self.sort)
-                result_index = (abs(key) - 1) % self.bulk_size
+                last_key = int(math.floor(
+                    float(self.count) / float(self.bulk_size)
+                )) * self.bulk_size
+                start = result_key = last_key - (
+                        (abs(key) / self.bulk_size) * self.bulk_size)
+                if last_key == result_key:
+                    result_index = key
+                else:
+                    result_index = (key % self.bulk_size) - (
+                        self.bulk_size - (self.count % last_key)
+                    )
 
             if result_key not in self.results:
                 self.results[result_key] = self.es._search(
-                    self.query, sort=sort, start=start, **self.query_params
-                )['hits']['hits']
+                    self.query, sort=self.sort, start=start,
+                    **self.query_params)['hits']['hits']
 
             return self.results[result_key][result_index]
 
@@ -135,7 +147,7 @@ class ElasticSearchCatalog(object):
             )
         return self._conn
 
-    def _search(self, query, **query_params):
+    def _search(self, query, sort=None, **query_params):
         '''
         '''
         if 'start' in query_params:
@@ -144,9 +156,13 @@ class ElasticSearchCatalog(object):
         query_params['fields'] = query_params.get('fields', 'path.path')
         query_params['size'] = self.get_setting('bulk_size', 50)
 
+        body = {'query': query}
+        if sort is not None:
+            body['sort'] = sort
+
         return self.connection.search(index=self.index_name,
                                       doc_type=self.doc_type,
-                                      body={'query': query},
+                                      body=body,
                                       **query_params)
 
     def search(self, query, factory=None, **query_params):
@@ -209,7 +225,6 @@ class ElasticSearchCatalog(object):
     def manage_catalogRebuild(self, *args, **kwargs):
         if self.enabled:
             self.recreateCatalog()
-
         return self.catalogtool._old_manage_catalogRebuild(*args, **kwargs)
 
     def manage_catalogClear(self, *args, **kwargs):