+++ collective/elasticsearch/indexes.py
@@ -147,7 +147,7 @@ class EDateIndex(BaseIndex):
             return {'range': {name: {'gte': first}}}
         elif range_ == 'max':
             return {'range': {name: {'lte': first}}}
-        elif range_ == 'minmax' and type(query) in (list, tuple) and \
+        elif range_ in ('minmax', 'min:max') and type(query) in (list, tuple) and \
                 len(query) == 2:
             return {
                 'and': [