+++ collective/elasticsearch/interfaces.py
@@ -29,6 +29,10 @@ class IAdditionalIndexDataProvider(Interface):
         pass
 
 
+class IReindexActive(Interface):
+    pass
+
+
 class IQueryAssembler(Interface):
     def normalize(query):
         pass